PRO-6812: tidied up exception handling
